create `build_helper/src/util` mod
